api/requirements.txt
Flask
Flask-Cors
flask-SQLAlchemy
psycopg2-binary
pytest
Flask-Migrate
Flask-Script
Flask-JWT-Extended
sqlathanor
stripe